requirements.txt
python-telegram-bot[asyncio]==20.6
python-dotenv
motor
